+++ src/lib.rs
@@ -129,39 +129,24 @@ impl Multiaddr {
         self
     }
 
-    /// Ensures the `Multiaddr` is a `/p2p/...` address for the given peer.
+    /// Parses a [`Multiaddr`] from the given string, ensuring that it ends with [`Protocol::P2p`] of the provided peer.
     ///
-    /// If the given address is already a `p2p` address for the given peer,
-    /// i.e. the last encapsulated protocol is `/p2p/<peer-id>`, this is a no-op.
+    /// If the address does not end with `/p2p`, it is simply appended.
+    /// If the address ends with a _different_ peer ID, parsing will fail.
+    pub fn from_str_and_peer(addr: &str, peer: PeerId) -> Result<Self> {
+        let addr = addr.parse::<Self>()?;
+        addr.with_p2p(peer).map_err(|_| Error::InvalidMultiaddr)
+    }
+
+    /// Appends the given [`PeerId`] if not yet present at the end of this multiaddress.
     ///
-    /// If the given address is already a `p2p` address for a different peer
-    /// than the one given, the peer-id is replaced with the one given.
-    ///
-    /// If the given address is not yet a `p2p` address,the `/p2p/<peer-id>`
-    /// protocol is appended to the returned address.
-    pub fn to_p2p_terminated(mut self, peer_id: PeerId) -> Self {
-        let mut last = None;
-        let mut index = 0;
-        loop {
-            let data = &self.bytes.as_ref()[index..];
-            if data.is_empty() {
-                break;
-            }
-
-            let (p, next_data) =
-                Protocol::from_bytes(data).expect("`Multiaddr` is known to be valid.");
-
-            last = Some((p, index));
-            index += data.len() - next_data.len();
-        }
-
-        match last {
-            Some((Protocol::P2p(p), _)) if p == peer_id => self,
-            Some((Protocol::P2p(_), index)) => {
-                Arc::make_mut(&mut self.bytes).truncate(index);
-                self.with(Protocol::P2p(peer_id))
-            }
-            _ => self.with(Protocol::P2p(peer_id)),
+    /// Fails if this address ends in a _different_ [`PeerId`].
+    /// In that case, the original, unmodified address is returned.
+    pub fn with_p2p(self, peer: PeerId) -> std::result::Result<Self, Self> {
+        match self.iter().last() {
+            Some(Protocol::P2p(p)) if p == peer => Ok(self),
+            Some(Protocol::P2p(_)) => Err(self),
+            _ => Ok(self.with(Protocol::P2p(peer))),
         }
     }